Better guard against showing the "new entry" button.
Its troubling me that if you start in your profile and go to another, the button would show up.  I'm not entirely sure how this is possible. As we completely destroy the profile component upon changing views and changing profiles.  Somehow this username was the active user's. Very troubling.  This fix seems to fix Kaley's problem, but I'm afraid that somehow she's seeing her blog on someone else's profile.

git-svn-id: https://repos.nextthought.com/svn/nti-svn/NextThoughtWebApp/trunk@17388 3dfb626b-079f-4ab1-a9f6-4654e5377508