HTML5 IndexedDB Data proxy research. (Will store javascript objects in a NoSQL database locally on the browser with Indexes and Keys, etc...)
This file is not referenced by the app code yet. To play with it, you must require it on the browser's repl. Ext.require('NextThought.proxy.IndexedDB') then declare a store to use it as it's proxy. See the commented out store declaration at the bottom of the file.

BTW... Its really fast. The 5300+ users on my system when dumped into it, read out in less than a second.

git-svn-id: https://repos.nextthought.com/svn/nti-svn/NextThoughtWebApp/trunk@32676 3dfb626b-079f-4ab1-a9f6-4654e5377508