check point. Filtering & searching students on an assignment now works.  Only thing left is to restyle the download button and re-add the 'request change' button.
This does not include counts for each filter. (open vs forcredit) We may not want to pay that cost up front. The UI elements have what they need and the styles are there, I'm just not filling in the counts. This could easily be a batchSize=1 for each filter but that seams wasteful for a small fluourish. The active filter reports the total, that should be good enough, no?

git-svn-id: https://repos.nextthought.com/svn/nti-svn/NextThoughtWebApp/trunk@34087 3dfb626b-079f-4ab1-a9f6-4654e5377508