NTI-1028 - O M F G ... this is embarrassing.
GUIDs can and DO start with numbers. Identifiers in HTML/CSS are not supposed to. (querySelector and querySelectorAll enforce this). This commit fixes the use of GUIDs to be prefixed with "guid-" so they always have a valid ID form. Second, this stops using a query to fetch the dom node whos ID we know. (we now get it by id)