NTI-1766 - unbreak navigating to assignments.
The "mimeType" field on models was not intended to be a constant... it was for the parser. It can be either a String or a String[]. We converted the mimeType from String to a String[]... which exposed bad assumptions of what .mimeType was. This fixes those assumptions where they broke for Assignment models.